refactor: delete capabilities projection shell
